test : PlayerMoney calculateResultMoneyTest 추가
calculateResultMoney의 계산식 수정 필요

+++ src/main/java/nextstep/blackjack/MoneyOperator.java
@@ -4,18 +4,18 @@ import java.math.BigDecimal;
 import java.util.function.BiFunction;
 
 public enum MoneyOperator {
-    PLUS((num1, num2) -> num2 + num1),
-    MINUS((num1, num2) -> num2 - num1),
+    PLUS((num1, num2) -> (int)(num2 + num1)),
+    MINUS((num1, num2) -> (int)(num2 - num1)),
     MULTIPLY((num1, num2)-> new BigDecimal(num2).multiply(new BigDecimal(num1)).intValue());
 
 //    private String operator;
-    private BiFunction<Integer, Integer, Integer> biFunction;
+    private BiFunction<Double, Integer, Integer> biFunction;
 
-    MoneyOperator(BiFunction<Integer, Integer, Integer> biFunction) {
+    MoneyOperator(BiFunction<Double, Integer, Integer> biFunction) {
         this.biFunction = biFunction;
     }
 
-    public Integer calculate(int num1, int num2) {
+    public Integer calculate(double num1, int num2) {
         return this.biFunction.apply(num1, num2);
